Move exception handlers and custom exceptions into the exceptions package Fix error handler to look for default error handlers using the class of the exception raised
