+++ mathesar/exceptions/error_codes.py
