+++ mathesar/tests/api/test_table_api.py
@@ -7,7 +7,7 @@ from sqlalchemy import text
 
 from mathesar import reflection
 from mathesar import models
-from mathesar.error_codes import ErrorCodes
+from mathesar.exceptions.error_codes import ErrorCodes
 from mathesar.models import Table, DataFile
 from db.tests.types import fixtures
 
@@ -463,7 +463,7 @@ def test_table_previews_wrong_column_number(client, schema, engine_email_type):
     response = client.post(f'/api/v0/tables/{table.id}/previews/', data=post_body)
     assert response.status_code == 400
     assert "number" in response.json()[0]['message']
-    assert ErrorCodes.ColumnSizeMismatch.value == response.json()[0]['error_code']
+    assert ErrorCodes.ColumnSizeMismatch.value == response.json()[0]['code']
 
 
 def test_table_previews_invalid_type_cast(client, schema, engine_email_type):
