+++ mathesar/tests/imports/test_csv.py
@@ -5,7 +5,7 @@ from sqlalchemy.exc import ProgrammingError
 from sqlalchemy import text
 
 from mathesar.models import DataFile, Schema
-from mathesar.errors import InvalidTableError
+from mathesar.exceptions.exceptions import InvalidTableError
 from mathesar.imports.csv import create_table_from_csv, get_sv_dialect, get_sv_reader
 from db.schemas.operations.create import create_schema
 from db.schemas.utils import get_schema_oid_from_name
